+++ NEWS.md
@@ -29,6 +29,7 @@
 * New writer named `write_mothur_taxonomy` to create an imitation of the mothur taxonomy format. 
 * New writer named `write_unite_general` to create an imitation of the UNITE general FASTA release. 
 * New writer named `write_silva_fasta` to create an imitation of the SILVA FASTA release. 
+* New function names `compare_treatments` to compare multiple samples in multiple treatments, applying a user-defined function.
 
 ##  metacoder 0.1.3